requirements.txt
Django
dj-database-url
django-dotenv
django-getenv
raven
uWSGI
newrelic